agent/requirements.txt
requests==2.32.3
python-dotenv==1.0.1
ollama==0.2.1
